MISRA: Register access improvements

+++ base/ail/vtss_util.h
@@ -29,7 +29,7 @@
 #define VTSS_BITMASK(x) ((1U << (x)) - 1)
 #endif
 #define VTSS_EXTRACT_BITFIELD(x, o, w) (((x) >> (o)) & VTSS_BITMASK(w))
-#define VTSS_ENCODE_BITFIELD(x, o, w)  (((x) & VTSS_BITMASK(w)) << (o))
+#define VTSS_ENCODE_BITFIELD(x, o, w)  (((u32)(x) & VTSS_BITMASK(w)) << (o))
 #define VTSS_ENCODE_BITMASK(o, w)      (VTSS_BITMASK(w) << (o))
 #define VTSS_BITOPS_DEFINED
 #endif /* VTSS_BITOPS_DEFINED */
